backend/full-yandex-sync/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
PyMuPDF==1.23.8
Pillow==10.2.0
